move update to its own module
