+++ eyes2-lib/src/entity/genotype/mod.rs
@@ -1,0 +1,2 @@
+mod genotype;
+mod random;
