+++ eyes2-lib/src/entity/mod.rs
@@ -1,4 +1,5 @@
 pub mod creature;
+pub mod update;
 
 pub use self::creature::Creature;
-pub use self::creature::{Update, UpdateQueue};
+pub use self::update::{Update, UpdateQueue};
